python-scripts/upload2remoteStorage/requirements.txt
oss2==2.15.0
ks3sdk==1.0.16
esdk_obs_python==3.21.8
boto3==1.18.41
botocore==1.21.41
obs==0.0.0
